WMLX——foreign-trade-email-sorter/foreign-trade-email-sorter/requirements.txt
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.0.0
